Quiz:submission|restriction

+++ database/migrations/2020_07_13_125359_create_quiz_submissions_table.php
@@ -17,7 +17,7 @@ class CreateQuizSubmissionsTable extends Migration
             $table->id();
             $table->integer('quiz_id');
             $table->integer('student_id');
-            $table->integer('marks');
+            $table->float('marks');
             $table->timestamps();
             $table->softDeletes()->nullable();
         });
